refactor: seperate api utils (grafana, web, etc)
